backend/agents/rag-agent/requirements.txt
# FastAPI 및 서버
fastapi==0.115.5
uvicorn[standard]==0.32.1

# LangChain 및 LangGraph
langchain>=0.3.10
langchain-openai>=0.2.10
langchain-community>=0.3.10
langchain-core>=0.3.21
langgraph>=0.2.53

# 벡터 스토어 및 임베딩
faiss-cpu>=1.9.0
openai==1.57.2

# PDF 처리
pypdf==5.1.0

# 텍스트 처리
langchain-text-splitters==0.3.2

# 환경 변수
python-dotenv==1.0.1

# 유틸리티
pydantic==2.10.3
pydantic-settings==2.6.1

